+++ src/proxy.js
@@ -1,5 +1,7 @@
 const http = require('http');
 
+const TIMEOUT = 10000;
+
 function proxyRequest(req, res, route) {
   const url = new URL(route.target);
   const options = {
@@ -8,7 +10,10 @@ function proxyRequest(req, res, route) {
     path: req.url,
     method: req.method,
     headers: { ...req.headers },
+    timeout: TIMEOUT,
   };
+  delete options.headers['host'];
+  delete options.headers['connection'];
   if (options.headers['cookie']) {
     const cookies = options.headers['cookie'].split('; ').filter(c => !c.startsWith('auth_session='));
     if (cookies.length) options.headers['cookie'] = cookies.join('; ');
@@ -20,8 +25,10 @@ function proxyRequest(req, res, route) {
     proxyRes.pipe(res);
   });
 
+  proxyReq.on('timeout', () => {
+    proxyReq.destroy(new Error('Proxy timeout'));
+  });
   proxyReq.on('error', err => {
-    console.error(`Proxy error [${route.description}]: ${err.message}`);
     if (!res.headersSent) {
       res.writeHead(502, { 'Content-Type': 'text/plain' });
       res.end(`Bad Gateway: ${route.description || route.target}`);
@@ -39,7 +46,10 @@ function proxyUpgrade(req, socket, head, route) {
     path: req.url,
     method: 'GET',
     headers: { ...req.headers },
+    timeout: TIMEOUT,
   };
+  delete options.headers['host'];
+  delete options.headers['connection'];
   if (options.headers['cookie']) {
     const cookies = options.headers['cookie'].split('; ').filter(c => !c.startsWith('auth_session='));
     if (cookies.length) options.headers['cookie'] = cookies.join('; ');
@@ -49,17 +59,15 @@ function proxyUpgrade(req, socket, head, route) {
   const proxyReq = http.request(options);
   proxyReq.on('upgrade', (proxyRes, proxySocket) => {
     socket.write('HTTP/1.1 101 Switching Protocols\r\n');
-    socket.write('Upgrade: websocket\r\n');
-    socket.write('Connection: Upgrade\r\n');
-    const accept = proxyRes.headers['sec-websocket-accept'];
-    if (accept) socket.write('Sec-WebSocket-Accept: ' + accept + '\r\n');
-    const proto = proxyRes.headers['sec-websocket-protocol'];
-    if (proto) socket.write('Sec-WebSocket-Protocol: ' + proto + '\r\n');
+    for (const [key, value] of Object.entries(proxyRes.headers)) {
+      socket.write(key + ': ' + value + '\r\n');
+    }
     socket.write('\r\n');
     socket.pipe(proxySocket);
     proxySocket.pipe(socket);
   });
-  proxyReq.on('error', () => socket.destroy());
+  proxyReq.on('timeout', () => proxyReq.destroy());
+  proxyReq.on('error', () => { try { socket.destroy(); } catch {} });
   proxyReq.end();
 }